Handle CJmp that can be either Guards or Branch
Due to incompleteness of DCFG, we may not be able to tell whether a CJmp is a in-loop branch or a guard until later Iterations.
As a workaround, we always add CJmp as a in-loop branch if we are not sure it is a guard, and remove it from branch list if it is.