Check constant dD before adding a Guard
dD for Guard was not checked, and as a result FuzzBALL may incorrectly create a loop summary even if dD is not constant (which means the Guard is not inductive variable and the loop is not feasible for summarization)
With this fix, FuzzBALL doesn not generate summary for such loops.
e.g. loop-acceleration/array_4 in svcomp